+++ Packages/Sources/TileKit/TileKit.Template.Rendering.swift
@@ -2,7 +2,7 @@ public extension TileKit.Template {
     protocol Rendering {
         func render(
             template: String,
-            context: [String: String],
+            context: Context,
         ) throws -> String
     }
 }